+++ board-server/src/main/java/net/stzups/board/config/StringConfig.java
@@ -1,3 +1,0 @@
-package net.stzups.board.config;
-
-public interface StringConfig extends Config<String> {}